db_tools.py
import mysql.connector
import os

def conn():
    return mysql.connector.connect(
        host="localhost",
        user="root",
        password=os.getenv("MYSQL_PASSWORD"),
        database="sakila"
    )