requirements.txt
mysql-connector-python
mcp
matplotlib
pandas